+++ parler/managers.py
@@ -33,7 +33,7 @@ class TranslatableQuerySet(QuerySet):
         Set the language code to assign to objects retrieved using this QuerySet.
         """
         if language_code is None:
-            language_code = appsettings.PARLER_DEFAULT_LANGUAGE_CODE
+            language_code = appsettings.PARLER_LANGUAGES.get_default_language()
 
         self._language = language_code
         return self
@@ -50,7 +50,7 @@ class TranslatableQuerySet(QuerySet):
 
         .. code-block:: python
 
-            qs.filter('en', name="Cheese Omelette")
+            qs.translated('en', name="Cheese Omelette")
 
         This will query the translated model for the ``name`` field.
         """
@@ -127,7 +127,7 @@ class TranslatableManager(models.Manager):
 
         .. code-block:: python
 
-            qs.filter('en', name="Cheese Omelette")
+            qs.translated('en', name="Cheese Omelette")
 
         This will query the translated model for the ``name`` field.
         """