+++ parler/models.py
@@ -12,12 +12,13 @@ or let it be created dynamically when using the :class:`TranslatedFields` field.
 """
 from __future__ import unicode_literals
 import django
+from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
 from django.db import models, router
 from django.db.models.base import ModelBase
 from django.db.models.fields.related import ReverseSingleRelatedObjectDescriptor
 from django.utils.functional import lazy
-from django.utils.translation import get_language, ugettext
+from django.utils.translation import get_language, ugettext, ugettext_lazy as _
 from django.utils import six
 from parler import signals
 from parler.cache import _cache_translation, _cache_translation_needs_fallback, _delete_cached_translation, get_cached_translation, _delete_cached_translations, get_cached_translated_field
@@ -372,9 +373,12 @@ class TranslatableModel(models.Model):
 
 
     def save_translations(self, *args, **kwargs):
+        # Copy cache, new objects (e.g. fallbacks) might be fetched if users override save_translation()
+        translations = self._translations_cache.values()
+
         # Save all translated objects which were fetched.
         # This also supports switching languages several times, and save everything in the end.
-        for translation in six.itervalues(self._translations_cache):
+        for translation in translations:
             if translation is None:  # Skip fallback markers
                 continue
 
@@ -498,7 +502,7 @@ class TranslatedFieldsModel(six.with_metaclass(TranslatedFieldsModelBase, models
     Base class for the model that holds the translated fields.
     """
 
-    language_code = models.CharField(max_length=15, db_index=True)
+    language_code = models.CharField(_("Language"), choices=settings.LANGUAGES, max_length=15, db_index=True)
     master = None   # FK to shared model.
 
     class Meta: